% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortical.R
\name{available_cortical_surfaces}
\alias{available_cortical_surfaces}
\title{List available cortical surfaces}
\usage{
available_cortical_surfaces()
}
\value{
Character vector of available surface names.
}
\description{
List available cortical surfaces
}
\examples{
available_cortical_surfaces()
}
\seealso{
Other cortical meshes:
\code{\link[=get_cortical_mesh]{get_cortical_mesh()}}
}
\concept{cortical meshes}
