% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebellar.R
\name{available_cerebellar_surfaces}
\alias{available_cerebellar_surfaces}
\title{List available cerebellar surfaces}
\usage{
available_cerebellar_surfaces()
}
\value{
Character vector of available surface names.
}
\description{
List available cerebellar surfaces
}
\examples{
available_cerebellar_surfaces()
}
