% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseg.meshes-package.R
\docType{package}
\name{ggseg.meshes-package}
\alias{ggseg.meshes}
\alias{ggseg.meshes-package}
\title{ggseg.meshes: Additional Brain Surface Meshes for the 'ggsegverse' Ecosystem}
\description{
Provides additional brain surface meshes for cortical and cerebellar visualisation in the 'ggseg' ecosystem. Cortical meshes include 'fsaverage5' pial, white, semi-inflated, sphere, smoothwm, and orig surfaces. Cerebellar meshes include the 'SUIT' flatmap surface. All meshes follow the same vertices/faces data frame format used by 'ggseg.formats' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggseg.meshes}
  \item \url{https://ggsegverse.github.io/ggseg.meshes/}
  \item Report bugs at \url{https://github.com/ggsegverse/ggseg.meshes/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
